@jwmerrill 2015-06-29
Graph of rendering time before (orange) and after (blue) this change.

![timing graph](https://s3.amazonaws.com/grapher/exports/sebigbmxk5.png)