@laughinghan 2015-06-30
I bet it would be fascinating to hear about how you isolated the perf problem to the `jQ.add()`; I guess this is a variant of the classic quadratic string accumulator problem.