@jwmerrill 2015-06-30
It was, perhaps surprisingly, not that hard to track this one down.

Evan Weinberg tweeted a graph at us that used to work, but started to not even load https://twitter.com/emwdx/status/615482918765334528. I just opened up Chrome's dev tools in the hung graph and hit the pause/play button a few times and noticed that every time I paused I was down in some jQuery sorting routine, and a few frames up the call stack was MathQuill's Fragment constructor. From there, like you suggest, it was a fairly obvious pattern match against other quadratic append situations I've found myself in.

I did end up spending a bit of time trying to understand why we didn't run into this same problem before. I looked through a bit of jQuery history, and there have been some changes in how often this sorting is enforced (used to be sometimes, now it's always), but somehow, the release version of mathquill doesn't seem to exhibit the issue, even with jQuery 2.1.4, and I'm not really sure why.